+++ RNFileUploadDownload/App/styles.ts
@@ -1,0 +1,9 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  rootContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    margin: 10,
+  },
+});
