+++ RNFileUploadDownload/src/components/Download/styles.ts
@@ -1,0 +1,16 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  buttonContainer: {
+    alignItems: 'flex-start',
+    borderWidth: 1,
+    borderRadius: 30,
+    padding: 15,
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: '600',
+    alignSelf: 'center',
+    padding: 10,
+  },
+});
