+++ RNFileUploadDownload/src/components/ProgressModal/types.ts
@@ -1,0 +1,8 @@
+import {GestureResponderEvent} from 'react-native';
+
+export type ProgressModalProps = {
+  visible: boolean;
+  onRequestClose: () => void;
+  preventParentEvent: (event: GestureResponderEvent) => void;
+  percent: number;
+};
