+++ RNFileUploadDownload/src/components/Upload/types.tsx
@@ -1,0 +1,12 @@
+export type OnInit = () => void;
+export type OnProgress = (percent: number) => void;
+export type OnEnd = () => void;
+
+export type ProgressProps = {
+  onInit: OnInit;
+  onProgress: OnProgress;
+  onEnd: OnEnd;
+};
+
+export type DownloadProps = ProgressProps & {link: string};
+export type UploadProps = ProgressProps;
