+++ RNFileUploadDownload/src/rn-fetch-blob/types.ts
@@ -1,0 +1,3 @@
+export type OnInit = () => void;
+export type OnProgress = (percent: number) => void;
+export type OnEnd = () => void;
